Fix high fraction liquidation security advisory

+++ packages/contracts/contracts/TestContracts/StabilityPoolTester.sol
@@ -10,7 +10,7 @@ contract StabilityPoolTester is StabilityPool {
         ETH = ETH.add(msg.value);
     }
 
-    function setCurrentScale(uint128 _currentScale) external {
+    function setCurrentScale(uint _currentScale) external {
         currentScale = _currentScale;
     }
